requirements.txt
fastapi>=0.112
uvicorn[standard]>=0.30
httpx>=0.27
requests>=2.32
Pillow>=10.3
python-multipart>=0.0.9
